Add static_len() on HList (Closes #59)
Since we know the length of HLists at compile time, there is no need
for a (&self) parameter. That said, we keep it around for the sake of
ergonomics and forward it to static_len().

len() is added, forwarding to static_len()

length() returning u32 is kept around, but deprecated. It also forwards to
static_len() now.