The Great Glob Import Extinction
This replaces almost all glob imports in doc examples,
examples/, and tests to use the more idiomatic, explicit
paths that #101 aimed to make possible.

`use` paths are now shown in the majority of doc tests,
and `extern crate` statements are visible in a couple of
doc tests that are sufficiently large or important
(e.g. the first example in the module's toplevel documentation)

There seemed to be little advantage to breaking this up, so I
just squashed it all together.
One thing worthy of note however is that frunk_core now has
'frunk' as a dev dependency, so that doc examples can use it.