Introduce Transmogrifying from A to B (#124)
This commit adds the ability to do "transmogrifying" (inspired by [this Haskell repo](https://github.com/ivan-m/transmogrify) and this [Scala SO answer](https://stackoverflow.com/a/51986872/1814775)) to Frunk.

What is "transmogrifying"? In this context, it means to convert some data of type A into data of type B, in a typesafe, recursive way, as long as A and B are "similarly-shaped".  In other words, as long as B's fields and their subfields are subsets of A's fields and their respective subfields, then A can be turned into B.