Optimized memory layout for Coproducts
Instead of nesting enums, the new representation explicitly has
one tag and nested unions. The nested unions only take up as much
space as the largest variant of the coproduct, unlike the old
representation, which was optimized very poorly by the compiler.

The code passes all tests but it is still quite messy and some
of the old Coproduct's API like Ord and Eq is still missing.